+++ src/main/java/de/tblsoft/solr/http/HTTPHelper.java
@@ -126,6 +126,9 @@ public class HTTPHelper {
 
 			responseBuilder.append(EntityUtils.toString(response.getEntity()));
 			httpclient.close();
+			if(response.getStatusLine().getStatusCode() >= 300) {
+				throw new RuntimeException("Problem httpCode: " + response.getStatusLine().getStatusCode() + " url: " + url + " payload: " + responseBuilder.toString());
+			}
 			return responseBuilder.toString();
 		} catch (Exception e) {
 			throw new RuntimeException(e);