+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes for the solr-cmd-utils project will be documented in this fi
 
 ## [Unreleased]
 ### Added
+
+
+### Changed
+
+
+### Fixed
+
+
+### Deprecated
+## [3.16] - 2021-02-20
+### Added
 - add a cache max age parameter to the http filter
 
 ### Changed